pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dealgraph"
version = "0.1.0"
description = "Graph-Augmented RAG Agent for private equity deal research"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "graph", "private-equity", "llm", "dspy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "faiss-cpu>=1.7",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "dspy-ai>=2.5",
    "typer>=0.9",
    "rich>=13.0",
    "tqdm>=4.65",
    "pytest>=9.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
dealgraph-agent = "dealgraph.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["dealgraph*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

